backend/requirements.txt
aiofiles==23.2.1
aiomysql==0.2.0
alembic==1.13.1
asyncpg==0.29.0
annotated-types==0.6.0
anyio==3.7.1
async-generator==1.10
attrs==23.2.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
fastapi-login==1.10.0
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
outcome==1.3.0.post0
passlib==1.7.4
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.7
PyYAML==6.0.1
setuptools==69.1.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.28
starlette==0.36.3
trio==0.21.0
typing_extensions==4.10.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
